Create Renderer service and start implementing it
There is a whole ton of code here. I have gutted PIXI.js
out of the engine because I'm sick of it. I have added
A renderer service to the services folder. Here I have
created most of the types of classes that we had in PIXI
except now we have much more control over the rendering
process. This includes rectangles, sprites, paths, and more
to come. Like in the droids-test branch, you will attach these
to you entity in the game plugin, but now you will invoke
the draw commands via the draw function attached to those
entities, (not implemented yet).

Drawing a rectangle given the world coordinate transformation
was super easy, so hopefully that is a good sign.