Add the ability to load and use sprites
Sprite can be rendered via the canvas renderer. There is also support
for loading texture via the AssetManager which are sprite sheets.
When any texture is loaded, an ajax query looks to see if there
is a corresponding json file with the same name as the sprite sheet.
Then the frame parameter of the sprite is used to determine what
square on the sprite sheet to render.