Updated some Documentation and added clear color
Since it's been so long since i've touched this code, I went through
and documented some of the functions I was writing in the renderer to
refamiliarize myself with the framework.

I also implemented the begin function in the canvas renderer to clear
to a clear color set at CanvasRenderer construction or at any time with
setClearColor.