Created a S-Expression scheme setup and improved parsing
In the initial droids-test branch using PIXI.JS, there was a little
trouble with tons and tons of boilerplate code to refine the
gamedata object into something more useful. I instead created a
list of string arrays containing the names of each member of each
S-Expression to make it easy to create the refined gamedata object.
Implementations of the Plugin must define the SexpScheme for
S-Expression files to parse correctly.

Also began the entity creation a little bit within the loadGame
function of the plugin. Next I will start using the new renderer.