Connected the currentTurn to the slider (finally!)
The slider never worked because it was never connected to the
Game services currentTurn member for some reason. Manually connected
it used broadcast messages from the Game service to be handled by the
controller to update the values within it's scope on changes from the
service.

Also slight tweaks to the PluginBase and Renderer. The new droids
example will not be pushed in this branch and will probably be included
farther down the line.