Option menus implemented
Created an Options service and a game-option directive to allow the
user to create a list of options to be modifiable by the user. Right
now, the only controls you can specify are a text field, a checkbox,
a slider, and a dropdown button. It is fairly easy to create more
types of controls accepted. Right now the arrays are implemented
in code, but later they can easily be moved to json scripts if that
seems desirable.