Updating the plugin system a bit
Needed to make it so that when plugins are loaded, it would clear
out the existing plugin and data, or at least clear out the data
to load a new game on the currently loaded plugin. Also made it so
that the engine looks for plugins inside their own folder within
the plugins folder, (goofy not to).

Also removed a bunch of voodoo printouts and moved some of the common
plugin functionality to it's base class.