Sprite rendering updated
Changed the way sprites are rendered to correctly allow for texture
coordinates to be applied to a drawing canvas. This allows subimages
within a texture to drawn to a canvas, like in most rendering API's.
Also added a tile offset x and y value for tiled texture to not
have to start at uv 0.0 in texture coordinates. Allows for cool texture
translation animations on tiled texture.