Starting work on transform matrices.
The plan is to have the main canvas, no matter the aspect ratio, will
be divided in units 100 x 100. Then individual transform matrices can
be attached to redering components, (Lines, Sprite, Rects, etc) which
will then be transformed with regards to the global matrix.If they don't
have a reference to a matrix, then their coordinates in going to be in
regards to the whole screen being 100 width and 100 height with origin
at the top left, positive x to the right, positive y downward.