Added drawRect and drawSprite to WebglRenderer
To do so I had to change how textures are loaded. Since textures
need to be loaded differently depending on the renderer, I removed
the AssetManager object altogether. Now the renderering itself is
responsible for the loading and unloading of its textures with a few
new added bits. Also, all matrices are now column major instead of
row major. Derped a little bit.