Added logic for selection
So selection works by taking the x,y coordinate relative to the
canvas, passing it to the Game object, which then calls the plugin
which hit tests as it pleases, returns a list of game data objects
which are passed to the navbar's debug table to be display with
angular magic. Way more work than i expected. (extra 3 or so hours)