Changed dependency naming, changed plugin manager
The dependency naming required that all the loadable modules created in
requirejs load from the /scripts directory and it was better for me to
more verbosely specify the location from root so that files not in the
scripts folder, (Like plugins) could be loaded with require.

Beside that, files that were edited inside the plugins folder did not
automatically spark the live recompile and livereload. That now happens.