Big changes to the way angular providers are created.
The way that the system was originally instantiating the services
was causing some duplication issues when requirejs became involved.
It wasn't that requirejs was instantiating the services multiple times
it's that every time the constructors were set for the specified service
before boostrapping occured, the services were then reinstantiated.

There was also a problem with invoked injected functions after
bootstrapping occurred, namely that apparently it's not possible.
I worked around this by making the injector for the module a public
member on the module, (like providers before), and then using classic
annotation injection for function invocations. Making everything
injected explicitly will allow to minify everything in the future which
has been a problem until now.