The minor angular update a little while back caused problems
Apparently they removed a controller syntax that i was used to using
where you add members to a controller, name the controller used in a
section of html code, and then access members like "playback.timedt"
where playback is the controller and timedt is the member. The other
way to do it is to attach the variables to the scope of the controller
with $scope and then in the body of the html code, you can access that
same variable like "timedt". Changed the playback bar to that new scheme
and it works like a charm.