Add Building Sprite
I meant to get this done earlier, but here we are.
Since we don't need to animate it, I went all out and put some nice
gradients on it to make it more interesting. Even if we did want to
animate it, I wouldn't need to redraw much or redo the gradients. If all
the other buildings had a similar gradient, it'd look pretty nice. I'll
put some info about the gradients on Slack or something.